src/HTTP_Trigger_app/requirements.txt
azure-functions
azure-search-documents
azure-core
langchain-core
langchain-openai
python-dotenv
openai
langchain-community
pypdf